h_functional: 修复100_int_literal和105_long_array2 (32→34/37) - lexer::parse_int: 用u32解析hex/oct再转i32，支持0x80000000等位模式 - emit_sw: 大偏移时用不同于rs的临时寄存器，避免clobber存储值 - emit_lw: 大偏移时用t0计算地址再加载到rd，避免rd被覆盖

+++ src/lexer.rs
@@ -295,9 +295,10 @@ impl<'a> Lexer<'a> {
             return Ok(TokenKind::IntLiteral(0));
         }
 
-        let value = i32::from_str_radix(digits, radix)
+        // Use u32 to allow hex literals like 0x80000000 (bit pattern for INT32_MIN)
+        let value = u32::from_str_radix(digits, radix)
             .map_err(|_| CompilerError::at(position, "integer literal out of range"))?;
-        Ok(TokenKind::IntLiteral(value))
+        Ok(TokenKind::IntLiteral(value as i32))
     }
 
     fn peek(&self) -> Option<u8> {
